requirements.txt
click==6.6
Flask==0.11
Flask-MySQL==1.3
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
ordereddict==1.1
python-dotenv==0.5.1
uWSGI==2.0.13.1
Werkzeug==0.11.10
